requirements.txt
psycopg2-binary
python-dotenv
sentence-transformers
faiss-cpu
numpy
llama-cpp-python
fastapi
uvicorn[standard]
